Fix: Split module home and video
