+++ app/build.gradle.kts
@@ -60,8 +60,8 @@ android {
 
 dependencies {
     // Add other module dependencies here
-    implementation(projects.feature.home.main)
-    implementation(projects.feature.home.detail)
+    implementation(projects.feature.home)
+    implementation(projects.feature.video)
     implementation(projects.feature.map)
     implementation(projects.feature.catalog)
     implementation(projects.feature.mountain)
