+++ app/src/main/java/jp/shirataki707/yamato/navigation/YamatoNavHost.kt
@@ -4,8 +4,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import jp.shirataki707.yamato.feature.catalog.navigation.catalogPage
-import jp.shirataki707.yamato.feature.home.detail.navigation.detailPage
-import jp.shirataki707.yamato.feature.home.detail.navigation.navigateToDetail
+import jp.shirataki707.yamato.feature.video.navigation.detailPage
+import jp.shirataki707.yamato.feature.video.navigation.navigateToDetail
 import jp.shirataki707.yamato.feature.home.main.navigation.HomeBaseRoute
 import jp.shirataki707.yamato.feature.home.main.navigation.homePage
 import jp.shirataki707.yamato.feature.map.navigation.mapPage
@@ -22,9 +22,7 @@ fun YamatoNavHost(
         startDestination = HomeBaseRoute,
         modifier = modifier,
     ) {
-        homePage(
-            onNavigateToDetail = navController::navigateToDetail,
-        ) {
+        homePage(onNavigateToDetail = navController::navigateToDetail) {
             detailPage(
                 onBackClick = navController::popBackStack,
             )
