--- feature/home/detail/src/main/kotlin/jp/shirataki707/yamato/feature/home/detail/ui/section/DetailContentSectionState.kt
@@ -1,3 +1,0 @@
-package jp.shirataki707.yamato.feature.home.detail.ui.section
-
-internal sealed interface DetailContentSectionState
