--- feature/home/detail/src/main/kotlin/jp/shirataki707/yamato/feature/home/detail/ui/section/DetailInitialSection.kt
@@ -1,20 +1,0 @@
-package jp.shirataki707.yamato.feature.home.detail.ui.section
-
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import jp.shirataki707.yamato.core.ui.common.CommonErrorContent
-import jp.shirataki707.yamato.core.ui.common.ParcelableResult
-
-@Composable
-internal fun DetailInitialSection(
-    sectionState: DetailInitialSectionState,
-    modifier: Modifier = Modifier,
-) {
-    if (sectionState.detailVideoResources is ParcelableResult.Failure) {
-        CommonErrorContent(
-            onReloadButtonClick = sectionState.initialLoadIfNeeded,
-            modifier = modifier.fillMaxSize(),
-        )
-    }
-}
