+++ feature/home/src/main/kotlin/jp/shirataki707/yamato/feature/home/ui/section/HomeLoadingSectionState.kt
