+++ feature/home_old/detail/.gitignore
@@ -1,0 +1,1 @@
+/build
