+++ feature/home_old/detail/src/main/AndroidManifest.xml
