+++ feature/video/src/main/kotlin/jp/shirataki707/yamato/feature/video/ui/section/DetailContentSectionState.kt
@@ -1,0 +1,3 @@
+package jp.shirataki707.yamato.feature.video.ui.section
+
+internal sealed interface DetailContentSectionState
