+++ feature/video/src/main/kotlin/jp/shirataki707/yamato/feature/video/ui/section/DetailLoadingSection.kt
@@ -1,0 +1,24 @@
+package jp.shirataki707.yamato.feature.video.ui.section
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+internal fun DetailLoadingSection(
+    sectionState: DetailLoadingSectionState,
+    modifier: Modifier = Modifier,
+) {
+    Box(
+        modifier = modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center,
+    ) {
+        CircularProgressIndicator(
+            color = MaterialTheme.colorScheme.primary,
+        )
+    }
+}
