+++ feature/video/src/main/kotlin/jp/shirataki707/yamato/feature/video/ui/section/DetailLoadingSectionState.kt
@@ -1,0 +1,13 @@
+package jp.shirataki707.yamato.feature.video.ui.section
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+
+internal data object DetailLoadingSectionState : DetailContentSectionState
+
+@Composable
+internal fun rememberDetailLoadingSectionState(): DetailLoadingSectionState {
+    return remember {
+        DetailLoadingSectionState
+    }
+}
