+++ settings.gradle.kts
@@ -29,7 +29,7 @@ include(":core:common")
 include(":core:network")
 include(":core:model")
 include(":core:network:youtube")
-include(":feature:home:main")
-include(":feature:home:detail")
 include(":feature:catalog")
 include(":feature:map")
+include(":feature:video")
+include(":feature:home")
